+++ ext/modules/payment/payu/complete.php
@@ -1,7 +1,7 @@
 <?php
 
 	/*
-		ver. 1.0.2
+		ver. 1.0.3
 		PayU Account Payment plugin for osCommerce 2.3.1
 		
 		@license    http://opensource.org/licenses/GPL-3.0  Open Software License (GPL 3.0)