+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.3
+
+* Removed weight value of product
+* Removed language code value
+* Removed SignatureTimestamp, Signature values
+
 ## 1.0.2
 
 * SDK 0.1.9.1 compatible